Implement read status synchronization
Controlled via the new defcustom `weechat-sync-buffer-read-status`, this
allows to mark buffers as read in the relay when they are read in the
client.

The idea was taken from weechat-android and "glowing bear", which all do
the same thing.

The feature can only be used if the relay version is >= 1.0.

Fixes #34