Don't store numbers in secrets.el attributes
It seems that the type of the attributes in  `secrets-create-item` must be strings. It is not apparent in [the documentation](https://www.gnu.org/software/emacs/manual/html_node/auth/Secret-Service-API.html). However, I was getting a `wrong type argument: stringp` error when using `weechat-secrets` (GNU Emacs 24.4.1).

This fixes the issue but since the attributes are apparently never retrieved, simply removing them may be considered a better option.